software/requirements.txt
numpy>=2.0.0
Pillow>=11.0.0
imageio>=2.30.0
imageio-ffmpeg>=0.5.0